+++ python/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enabled session management via `EndSession`  [#1227]
 - Support for temporary or permanent `safety-checks` setting
 - Support for Output Descriptors export [#1363]
+- PIN entry via letters  [#1496]
 
 ### Changed
 
@@ -514,3 +515,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1232]: https://github.com/trezor/trezor-firmware/issues/1232
 [#1266]: https://github.com/trezor/trezor-firmware/issues/1266
 [#1363]: https://github.com/trezor/trezor-firmware/pull/1363
+[#1496]: https://github.com/trezor/trezor-firmware/pull/1496